feat: Major platform overhaul - JWT auth, 20 strategies, dark theme, advanced backtesting
- Backend: JWT authentication (register/login), CORS fix, async backtesting,
  yfinance caching, N+1 query fixes, bulk inserts, new market endpoints
- Database: 15 models (8 new: Portfolio, Watchlist, Alert, MarketDataCache,
  UserSettings, AuditLog, StrategyPerformance, OrderHistory), Alembic migrations,
  Integer->Float for financial fields, proper indexes/constraints/enums
- Frontend: Dark theme (TradingView-style), 5 chart components (candlestick,
  equity curve, drawdown, win/loss, radar), 9 pages (Login, Portfolio,
  TradeJournal, RiskAnalytics, Settings), JWT auth flow, toast notifications
- Strategies: 10 new (VWAP, KAMA, MeanReversion, OrderFlow, VolumeProfile,
  MultiTF, GARCH, Wyckoff, PairsTrading, Composite), bugs fixed in existing 10
  (AI look-ahead bias, SL/TP order, missing validations, Order Blocks)
- Backtesting: Entry at next open, Sharpe/Sortino/Calmar metrics, walk-forward
  analysis, Monte Carlo simulation, signal_strength position sizing
- Tests: 14 tests (health + strategy CRUD + auth)
- Docs: Updated README.md and CLAUDE.md with full project documentation

+++ backend/app/schemas/strategy.py
@@ -8,9 +8,9 @@ class StrategyBase(BaseModel):
     description: Optional[str] = None
     strategy_type: str
     config: Dict[str, Any]
-    initial_capital: Optional[int] = 10000
-    stop_loss_pct: Optional[int] = 2
-    take_profit_rr: Optional[int] = 2
+    initial_capital: Optional[float] = 10000.0
+    stop_loss_pct: Optional[float] = 2.0
+    take_profit_rr: Optional[float] = 2.0
 
 
 class StrategyCreate(StrategyBase):
@@ -37,6 +37,7 @@ class Strategy(StrategyBase):
 
 
 from .backtest import BacktestRun
+
 
 class StrategyWithBacktests(Strategy):
     backtest_runs: List[BacktestRun] = []
